Clean up for PR
